==> 0_Demographics.Rmd <==
---
title: "Demographics"
output: 
  html_document

---

<!-- Set general settings -->

```{r setup, include = FALSE}

# Set general settings for markdown file
knitr::opts_chunk$set(
  message = FALSE,
  warning = FALSE,
  comment = "",
  results = "hold"
)


# Clear environment
rm(list = ls())


# Enable/disable caching of time-consuming code chunks
knitr_cache_enabled = TRUE


# Load packages
library(dplyr)      # for data manipulation
library(knitr)      # for integrating computing and reporting in markdown
library(kableExtra) # for customizing appearance of tables
library(foreign)    # for reading in SPSS file
library(tidyr)      # for gather and spread function
library(lubridate)  # to convert seconds to minutes and seconds


# Load functions
source("./functions/my_table_template.R") # Function to create table template


# Turn off scientific notation
options(scipen = 999)
```
<br><br> 

## Sample Overview
***

* N = 56 (N HC = 28, N OCD = 28)

* OCD and HC matched for gender, age, education level, and handedness 

* Native German speakers with normal or corrected-to-normal vision

* Patients with OCD were recruited from the outpatient clinic at Humboldt-Universität zu Berlin. All patients were interviewed and diagnosed by trained psychologists using the Structured Clinical Interview for DSM-IV (SCID-I; Wittchen et al., 1997) and met the DSM-IV criteria for OCD as primary disorder. 

* Exclusion criteria 
  - general: history of any neurological disorder, current or lifetime diagnosis of a substance-related disorder, schizophrenia spectrum disorder, bipolar disorder, use of benzodiazepines during the last week or of neuroleptic medication during last three months, and general exclusion criteria for tDCS (e.g., any metal or electronic implants in the head or upper body, skin disorder or skin condition at or near stimulation locations, pregnancy)
  - HC: self-reported history of any current or past axis I disorder or psychotherapeutic treatment

```{r demographics}

# Load data 
participant_data <- read.spss("./data/Probandendaten.sav", to.data.frame = TRUE)
questionnaires   <- read.spss("./data/psychoEQExport.sav", to.data.frame = TRUE, add.undeclared.levels = "no") # add.undeclared.levels required to prevent irrelevant warning message


# Select relevant variables
participant_data <- participant_data[,c("Code","Alter","Geschlecht","Schulabschluss","Schuljahre","Händigkeit","Komorbidität_1","Komorbidität_2",
                                        "Komorbidität_3","Medikamente_JaNein","SSRI","SNRI","Trizykl_Antidepressiva",
                                        "Kombination_SSRI_Antidepressiva","YBOCS_gesamt","YBOCS_Gedanken","YBOCS_Handlungen","MADRS", "WST",
                                        "Dauer_nach_tDCS_bis_Flanker_T1","Dauer_nach_tDCS_bis_Flanker_T2","Tage_zw_Sessions",
                                        "Jucken_T1","Kribbeln_T1","Brennen_T1","Wärme_Hitze_T1","Schmerz_T1","Kopfschmerzen_T1", "Ermüdung_verringerte_Aufmerksamkeit_T1",
                                        "Konzentrationsschwierigkeit_T1","Stimmungsänderung_T1","Nervosität_T1","sonstiges_T1","Zeitpunkt_Empfindung_T1",
                                        "Auswirkung_Empfindung_auf_Leistung_T1","Lichtblitze_T1", "Vermutete_Stimulationsart_T1"
                                        ,"Jucken_T2","Kribbeln_T2","Brennen_T2","Wärme_Hitze_T2","Schmerz_T2","Kopfschmerzen_T2", "Ermüdung_verringerte_Aufmerksamkeit_T2",
                                        "Konzentrationsschwierigkeit_T2","Stimmungsänderung_T2","Nervosität_T2","sonstiges_T2","Zeitpunkt_Empfindung_T2",
                                        "Auswirkung_Empfindung_auf_Leistung_T2","Lichtblitze_T2", "Vermutete_Stimulationsart_T2",
                                        "Stimulationsbedingung_T1", "Stimulationsbedingung_T2")]
questionnaires   <- questionnaires[,c("CODE", "BD2SUMT0", "OCISUMT0", "STSSUMT0", "STTSUMT0")] 


# Merge dataframes and rename columns
participant_data$Code <- gsub(" ", "", participant_data$Code , fixed = TRUE) # delete empty spaces from string to enable merging
demographics <- left_join(participant_data, questionnaires, by = c("Code" = "CODE")) %>%
  # Rename some columns
  dplyr::rename(participant_id = Code,
                BDI_Total      = BD2SUMT0, 
                OCI_Total      = OCISUMT0, 
                STAI_State     = STSSUMT0, 
                STAI_Trait     = STTSUMT0) %>%
  # Add group variable
  dplyr::mutate(group = as.factor(ifelse(substr(participant_id,13,13) == "C", "HC", "OCD")))   


# Exclude P_02 (due to retainer) and C_02 (as preregistered: patients are excluded with their match)
demographics <- demographics[demographics$participant_id != "ModERN_tDCS_P02" & demographics$participant_id != "ModERN_tDCS_C02",]


# C_12 repeated the WST at T2, because score at T1 was very low (22); score achieved at T2 was 25; we will use this score 
demographics[demographics$participant_id == "ModERN_tDCS_C12",]$WST <- 25



### Handle numeric variables

# Transform WST scores (range 24-37) to IQ scores
demographics[demographics$WST == 24,]$WST <- 89
demographics[demographics$WST == 25,]$WST <- 90
demographics[demographics$WST == 27,]$WST <- 93
demographics[demographics$WST == 28,]$WST <- 95
demographics[demographics$WST == 29,]$WST <- 97
demographics[demographics$WST == 30,]$WST <- 99
demographics[demographics$WST == 31,]$WST <- 101
demographics[demographics$WST == 32,]$WST <- 104
demographics[demographics$WST == 33,]$WST <- 107
demographics[demographics$WST == 34,]$WST <- 110
demographics[demographics$WST == 35,]$WST <- 114
demographics[demographics$WST == 36,]$WST <- 118
demographics[demographics$WST == 37,]$WST <- 122


# Calculate M and SD per group 
demographics_per_group_M_SD <- demographics %>%
  dplyr::group_by(group) %>%
  # Calculate M and SD of the variables
  dplyr::summarize(across(c(Alter,Schuljahre,WST,YBOCS_gesamt,YBOCS_Gedanken,YBOCS_Handlungen,MADRS,
                            BDI_Total,OCI_Total,STAI_State,STAI_Trait), list(mean,sd), na.rm = TRUE)) %>%
  # Reshape table to have one column per group
  gather(key = key, value = value, 2:ncol(.), factor_key = TRUE) %>% 
  spread(key = names(.)[1], value = "value") %>%
  # Add variable column
  dplyr::mutate(Variable = c("Alter", "","Schuljahre", "", "WST", "", "YBOCS_Total", "", "YBOCS_Gedanken", "", "YBOCS_Handlungen", "",  "MADRS", "", "BDI_Total", "", "OCI_Total", "", "STAI_State", "", "STAI_Trait", "")) %>%
  # Add M/SD column
  dplyr::mutate(M_SD_or_n = c("M", "SD","M", "SD", "M", "SD", "M", "SD", "M", "SD", "M", "SD", "M", "SD", "M", "SD", "M", "SD", "M", "SD", "M", "SD")) %>%
  # Reorder columns
  dplyr::select(Variable, M_SD_or_n, HC, OCD) %>%
  dplyr::ungroup()


# Calculate t tests
demographics_per_group_t_test <- demographics %>% 
  select(Alter,Schuljahre,WST,BDI_Total,OCI_Total,STAI_State,STAI_Trait, group) %>% 
  gather(key = Variable, value = value, -group) %>% 
  group_by(group, Variable) %>% 
  dplyr::summarise(value = list(value)) %>% 
  spread(group, value) %>%
  dplyr::ungroup() %>%
  dplyr::group_by(Variable) %>% 
  dplyr::mutate(df = t.test(unlist(HC), unlist(OCD), var.equal=TRUE)$parameter,
         t_or_chi_sq_value = t.test(unlist(HC), unlist(OCD), var.equal=TRUE)$statistic, 
         p_value = t.test(unlist(HC), unlist(OCD), var.equal=TRUE)$p.value) %>%
  dplyr::select(-c(HC, OCD))  %>%
  dplyr::ungroup()


# Merge M_SD and t test infos
demographics_per_group <- left_join(demographics_per_group_M_SD, demographics_per_group_t_test, by = "Variable")




### Handle count variables

# Calculate Chi-squared tests
gender_test     <- chisq.test(demographics$Geschlecht, demographics$group)
education_test  <- chisq.test(demographics$Schulabschluss, demographics$group)
handedness_test <- chisq.test(demographics$"Händigkeit", demographics$group)


# Extract counts for gender
gender <- demographics %>% group_by(group) %>% count(Geschlecht) %>% 
  # Reshape table to have one column per group
  spread(key = names(.)[1], value = n) %>%
  # Add chi square results
  dplyr::mutate(df = ifelse(row_number()==1, gender_test$parameter, NA),
                t_or_chi_sq_value = ifelse(row_number()==1, gender_test$statistic, NA),
                p_value = ifelse(row_number()==1, gender_test$p.value, NA)) %>%
  # Add first column
  dplyr::mutate(Variable = c("Geschlecht", "")) %>%
  # Select and reorder columns
  dplyr::select(Variable, everything()) %>%
  # Rename one column to make rbind wirh columns above possible
  dplyr::rename(M_SD_or_n = Geschlecht)  %>%
  dplyr::ungroup()


# Extract counts for education
education <- demographics %>% group_by(group) %>% count(Schulabschluss, .drop = FALSE) %>% 
  # Reshape table to have one column per group
  spread(key = names(.)[1], value = n) %>%
  # Add chi square results
  dplyr::mutate(df = ifelse(row_number()==1, education_test$parameter, NA),
                t_or_chi_sq_value = ifelse(row_number()==1, education_test$statistic, NA),
                p_value = ifelse(row_number()==1, education_test$p.value, NA)) %>%
  # Add first column
  dplyr::mutate(Variable = c("Schulabschluss", "", "" , "", "" , "")) %>%
  # Select and reorder columns
  dplyr::select(Variable, everything()) %>%
  # Rename one column to make rbind wirh columns above possible
  dplyr::rename(M_SD_or_n = Schulabschluss)  %>%
  dplyr::ungroup()


# Extract counts for handedness
handedness <- demographics %>% group_by(group) %>% count(Händigkeit, .drop = FALSE) %>% 
  # Reshape table to have one column per group
  spread(key = names(.)[1], value = n) %>%
  # Add chi square results
  dplyr::mutate(df = ifelse(row_number()==1, handedness_test$parameter, NA),
                t_or_chi_sq_value = ifelse(row_number()==1, handedness_test$statistic, NA),
                p_value = ifelse(row_number()==1, handedness_test$p.value, NA)) %>%
  # Add first column
  dplyr::mutate(Variable = c("Händigkeit", "", "")) %>%
  # Select and reorder columns
  dplyr::select(Variable, everything()) %>%
  # Rename one column to make rbind wirh columns above possible
  dplyr::rename(M_SD_or_n = Händigkeit)  %>%
  dplyr::ungroup()



# Display table
my_table_template(rbind(demographics_per_group[c(1,2),], gender, education, handedness, demographics_per_group[c(3:nrow(demographics_per_group)),]), caption = "Demographic and Clinical Variables", digits = 2)

````
<br><br>

## Overview tDCS time intervals

```{r tDCS-time-intervals}

# Create df with duration between stimulation and task in one column
time <- demographics %>%
  select(Dauer_nach_tDCS_bis_Flanker_T1,Dauer_nach_tDCS_bis_Flanker_T2) %>%
  gather(key = Variable, value = seconds)

interval_mean <- seconds_to_period(mean(time$seconds, na.rm = TRUE))
interval_sd   <- seconds_to_period(sd(time$seconds, na.rm = TRUE))
interval_min  <- seconds_to_period(min(time$seconds, na.rm = TRUE))
interval_max  <- seconds_to_period(max(time$seconds, na.rm = TRUE))
```

Days between sessions: *M* = `r round(mean(demographics$Tage_zw_Sessions, na.rm = TRUE), digits = 2)` days, *SD* = `r round(sd(demographics$Tage_zw_Sessions, na.rm = TRUE), digits = 2)`, *range* = `r min(demographics$Tage_zw_Sessions)`- `r max(demographics$Tage_zw_Sessions)`.

Duration between stimulation and task: *M* = `r interval_mean$minute`:`r round(interval_mean$second,digits=0)` min, *SD* = `r interval_sd$minute`:`r round(interval_sd$second,digits=0)`, *range* = `r interval_min$minute`:`r round(interval_min$second,digits=0)`- `r interval_max$minute`:`r round(interval_max$second,digits=0)`.
<br><br><br><br>

## Overview tDCS adverse effects

```{r tDCS-adverse-effects}

# Create overview adverse effects
table_adverse_effects <- demographics %>% 
  # Select the needed columns
  dplyr::select(participant_id, Jucken_T1, Jucken_T2, Kribbeln_T1, Kribbeln_T2, Brennen_T1, Brennen_T2, Wärme_Hitze_T1, Wärme_Hitze_T2,
         Schmerz_T1, Schmerz_T2, Kopfschmerzen_T1, Kopfschmerzen_T2, Ermüdung_verringerte_Aufmerksamkeit_T1, Ermüdung_verringerte_Aufmerksamkeit_T2, 
         Konzentrationsschwierigkeit_T1, Konzentrationsschwierigkeit_T2, Stimmungsänderung_T1, Stimmungsänderung_T2, 
         Nervosität_T1, Nervosität_T2, Auswirkung_Empfindung_auf_Leistung_T1, Auswirkung_Empfindung_auf_Leistung_T2,
         Lichtblitze_T1, Lichtblitze_T2, Vermutete_Stimulationsart_T1, Vermutete_Stimulationsart_T2, Stimulationsbedingung_T1, Stimulationsbedingung_T2) %>%
  # Stack T1 and T2 for each adverse effect below each other (this also stacks all adverse effects in one column)
  gather(adverse_effect, value, Jucken_T1:Jucken_T2, Kribbeln_T1:Kribbeln_T2, Brennen_T1:Brennen_T2, Wärme_Hitze_T1:Wärme_Hitze_T2,
         Schmerz_T1:Schmerz_T2, Kopfschmerzen_T1:Kopfschmerzen_T2, Ermüdung_verringerte_Aufmerksamkeit_T1:Ermüdung_verringerte_Aufmerksamkeit_T2, 
         Konzentrationsschwierigkeit_T1:Konzentrationsschwierigkeit_T2, Stimmungsänderung_T1:Stimmungsänderung_T2, 
         Nervosität_T1:Nervosität_T2, Auswirkung_Empfindung_auf_Leistung_T1:Auswirkung_Empfindung_auf_Leistung_T2,
         Lichtblitze_T1:Lichtblitze_T2, Vermutete_Stimulationsart_T1:Vermutete_Stimulationsart_T2, Stimulationsbedingung_T1:Stimulationsbedingung_T2) %>%
  # Create column for session and remove session from name of adverse effect
  dplyr::mutate(session = substr(adverse_effect, nchar(adverse_effect) - 2 + 1, nchar(adverse_effect)), # extract substring of last 2 characters of adverse_effect
         adverse_effect = substr(adverse_effect,1,nchar(adverse_effect)-3)) %>% # remove substring of last 3 characters of adverse_effect
  # Reshape table to show not all adverse effects  below each other in one column but have one column for each adverse effect
  spread(key = adverse_effect, value = value) 


# Create and display contingency tables
my_table_template(prop.table(table(table_adverse_effects$Jucken, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Jucken (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Kribbeln, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Kribbeln (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Brennen, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Brennen (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Wärme_Hitze, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Wärme_Hitze (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Schmerz, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Schmerz (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Kopfschmerzen, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Kopfschmerzen (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Ermüdung_verringerte_Aufmerksamkeit, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Ermüdung_verringerte_Aufmerksamkeit (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Konzentrationsschwierigkeit, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Konzentrationsschwierigkeit (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Stimmungsänderung, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Stimmungsänderung (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Nervosität, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Nervosität (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Auswirkung_Empfindung_auf_Leistung, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Auswirkung_Empfindung_auf_Leistung (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Lichtblitze, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Lichtblitze (%)", row_names = TRUE)
my_table_template(prop.table(table(table_adverse_effects$Vermutete_Stimulationsart, table_adverse_effects$Stimulationsbedingung), margin=2)*100, caption = "Vermutete_Stimulationsart (%)", row_names = TRUE)


# Calculate chi-squared tests
chisq.test(table_adverse_effects$Jucken, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Kribbeln, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Brennen, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Wärme_Hitze, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Schmerz, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Kopfschmerzen, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Ermüdung_verringerte_Aufmerksamkeit, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Konzentrationsschwierigkeit, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Stimmungsänderung, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Nervosität, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Auswirkung_Empfindung_auf_Leistung, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Lichtblitze, table_adverse_effects$Stimulationsbedingung)
chisq.test(table_adverse_effects$Vermutete_Stimulationsart, table_adverse_effects$Stimulationsbedingung)


# # Calculate Fisher's exact tests
# fisher.test(table_adverse_effects$Jucken, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Kribbeln, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Brennen, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Wärme_Hitze, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Schmerz, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Kopfschmerzen, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Ermüdung_verringerte_Aufmerksamkeit, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Konzentrationsschwierigkeit, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Stimmungsänderung, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Nervosität, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Auswirkung_Empfindung_auf_Leistung, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Lichtblitze, table_adverse_effects$Stimulationsbedingung)
# fisher.test(table_adverse_effects$Vermutete_Stimulationsart, table_adverse_effects$Stimulationsbedingung)



#### Double check: Treat rating as interval data and run t tests

# Recode variables to numeric values
table_adverse_effects_num <- table_adverse_effects 
table_adverse_effects_num[table_adverse_effects_num=="keine"]  <- 0
table_adverse_effects_num[table_adverse_effects_num=="mild"]   <- 1
table_adverse_effects_num[table_adverse_effects_num=="moderat"]<- 2
table_adverse_effects_num[table_adverse_effects_num=="stark"]  <- 3
table_adverse_effects_num[table_adverse_effects_num=="überhaupt nicht"]  <- 0
table_adverse_effects_num[table_adverse_effects_num=="ein wenig"]        <- 1
table_adverse_effects_num[table_adverse_effects_num=="mäßig"]            <- 2
table_adverse_effects_num[table_adverse_effects_num=="sehr stark"]       <- 4
table_adverse_effects_num[table_adverse_effects_num=="ja"]    <- 1


# Make column type numeric 
table_adverse_effects_num[, c(3:13,16)] <- sapply(table_adverse_effects_num[, c(3:13,16)], as.numeric)


# Calculate means
table_adverse_effects_num_means <- table_adverse_effects_num %>%
  dplyr::group_by(Stimulationsbedingung) %>%
  dplyr::summarise_if(is.numeric, funs(mean,sd), na.rm = TRUE) %>%
  dplyr::ungroup()


# Display means
my_table_template(table_adverse_effects_num_means, caption = "Mean of adverse effects, treated as interval data")


# Calculate t tests (not adjusted for multiple testing)
t.test(Jucken ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Kribbeln ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Brennen ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Wärme_Hitze ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE, na.action = "na.pass") # exclusion of participant with missing value (P_03 excluded, due to missing value in T1)
t.test(Schmerz ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE, na.action = "na.pass") # exclusion of participant with missing value (P_01 excluded, due to missing value in T2)
t.test(Kopfschmerzen ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Ermüdung_verringerte_Aufmerksamkeit ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Konzentrationsschwierigkeit ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Stimmungsänderung ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Nervosität ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Auswirkung_Empfindung_auf_Leistung ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)
t.test(Lichtblitze ~ Stimulationsbedingung, table_adverse_effects_num, paired = TRUE)


# Calculate t tests (adjusted for multiple testing)
t1 <- pairwise.t.test(table_adverse_effects_num$Jucken, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t2 <- pairwise.t.test(table_adverse_effects_num$Kribbeln, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t3 <- pairwise.t.test(table_adverse_effects_num$Brennen, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t4 <- pairwise.t.test(table_adverse_effects_num$Wärme_Hitze, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE, na.action = "na.pass") # exclusion of participant with missing value (P_03 excluded, due to missing value in T1)
t5 <- pairwise.t.test(table_adverse_effects_num$Schmerz, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE, na.action = "na.pass") # exclusion of participant with missing value (P_01 excluded, due to missing value in T2)
t6 <- pairwise.t.test(table_adverse_effects_num$Kopfschmerzen, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t7 <- pairwise.t.test(table_adverse_effects_num$Ermüdung_verringerte_Aufmerksamkeit, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t8 <- pairwise.t.test(table_adverse_effects_num$Konzentrationsschwierigkeit, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t9 <- pairwise.t.test(table_adverse_effects_num$Stimmungsänderung, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t10 <- pairwise.t.test(table_adverse_effects_num$Nervosität, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t11 <- pairwise.t.test(table_adverse_effects_num$Auswirkung_Empfindung_auf_Leistung, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)
t12 <- pairwise.t.test(table_adverse_effects_num$Lichtblitze, table_adverse_effects_num$Stimulationsbedingung, paired = TRUE)


# Adjust p values for multiple testing
print("# False discovery rate adjusted p values")
p_values <- c(t1$p.value,t2$p.value,t3$p.value,t4$p.value,t5$p.value,t6$p.value,t7$p.value,t8$p.value,t9$p.value,t10$p.value)
p.adjust(p_values, method = "fdr")
```
<br><br>

The stimulation was well tolerated with minimal adverse effects. The most frequent adverse effects were transient tingling, itching, and burning sensation. Adverse effects were not significantly different between verum and sham stimulation. (This holds true when evaluating rating as ordinal data with chi-square test or as interval data with t test). Also, participants were indeed blind to stimulation condition.
