==> 3_ERP_Visualization.Rmd <==
---
title: "ERP Visualization"
output: 
  html_document

---

<!-- Set general settings -->

```{r setup, include = FALSE}

# Set general settings for markdown file
knitr::opts_chunk$set(
  message = FALSE,
  warning = FALSE,
  comment = "",
  results = "hold"
)


# Clear environment
rm(list = ls())


# Enable/disable caching of time-consuming code chunks
knitr_cache_enabled = TRUE


# Load packages
library(dplyr)      # for data manipulation
library(ggplot2)    # for plotting
library(cowplot)    # for arranging plots
library(purrr)      # for calculate within-participant CIs for ERP trajectories
library(eegUtils)   # for plotting topographies
library(tidyr)      # for gather function 


# Load functions
source("./functions/summarySEwithinO.R")  # Function provided by R-cookbook: http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/


# Set figure theme and colors
my_figure_theme <- theme_classic(base_size = 11) +
  theme(legend.position = "bottom", 
        strip.background = element_rect(fill="grey95", linetype = "blank"),
        plot.title = element_text(hjust = 0.5),
        panel.border = element_rect(colour = "grey", fill=NA, size=.5), 
        panel.grid.major.y = element_line(colour = "grey90", linetype = "dotted", size=0.3))
        

# instad of theme_classic: + theme_apa(base_size = 11)

my_figure_colors <- c("slategray3","navy","tan1","sienna3")
```

## Response-locked ERPs
***

<!-- Load and clean data -->

```{r load-and-clean-data-resp-locked}

# Load data
resp_locked_data <- read.csv(file = "./data/response_locked_data_for_plots.csv", header = TRUE)


# Rename conditions
resp_locked_data$condition[resp_locked_data$condition == 1] <- "correct";
resp_locked_data$condition[resp_locked_data$condition == 2] <- "incorrect";



# Add stimulation condition variable
resp_locked_data$stimulation <- NA
resp_locked_data[
  resp_locked_data$participant_id =="C_01_T1"| resp_locked_data$participant_id =="C_02_T1"| resp_locked_data$participant_id =="C_03_T2"|
  resp_locked_data$participant_id =="C_04_T2"| resp_locked_data$participant_id =="C_05_T1"| resp_locked_data$participant_id =="C_06_T2"|
  resp_locked_data$participant_id =="C_07_T1"| resp_locked_data$participant_id =="C_08_T1"| resp_locked_data$participant_id =="C_09_T2"|
  resp_locked_data$participant_id =="C_10_T2"| resp_locked_data$participant_id =="C_11_T2"| resp_locked_data$participant_id =="C_12_T2"|
  resp_locked_data$participant_id =="C_13_T1"| resp_locked_data$participant_id =="C_15_T2"| resp_locked_data$participant_id =="C_16_T1"|
  resp_locked_data$participant_id =="C_17_T2"| resp_locked_data$participant_id =="C_18_T2"| resp_locked_data$participant_id =="C_19_T1"|
  resp_locked_data$participant_id =="C_20_T1"| resp_locked_data$participant_id =="C_21_T2"| resp_locked_data$participant_id =="C_22_T2"|
  resp_locked_data$participant_id =="C_23_T1"| resp_locked_data$participant_id =="C_24_T1"| resp_locked_data$participant_id =="C_25_T1"|
  resp_locked_data$participant_id =="C_26_T1"| resp_locked_data$participant_id =="C_27_T2"| resp_locked_data$participant_id =="C_28_T2"|
  resp_locked_data$participant_id =="C_29_T1"| resp_locked_data$participant_id =="C_30_T2"| resp_locked_data$participant_id =="P_01_T1"|
  resp_locked_data$participant_id =="P_02_T2"| resp_locked_data$participant_id =="P_03_T2"| resp_locked_data$participant_id =="P_04_T1"|
  resp_locked_data$participant_id =="P_05_T1"| resp_locked_data$participant_id =="P_06_T1"| resp_locked_data$participant_id =="P_07_T2"|
  resp_locked_data$participant_id =="P_08_T2"| resp_locked_data$participant_id =="P_09_T2"| resp_locked_data$participant_id =="P_10_T1"|
  resp_locked_data$participant_id =="P_11_T2"| resp_locked_data$participant_id =="P_12_T2"| resp_locked_data$participant_id =="P_13_T1"|
  resp_locked_data$participant_id =="P_15_T1"| resp_locked_data$participant_id =="P_16_T1"| resp_locked_data$participant_id =="P_17_T2"|
  resp_locked_data$participant_id =="P_18_T1"| resp_locked_data$participant_id =="P_19_T1"| resp_locked_data$participant_id =="P_20_T2"|
  resp_locked_data$participant_id =="P_21_T2"| resp_locked_data$participant_id =="P_22_T1"| resp_locked_data$participant_id =="P_23_T1"|
  resp_locked_data$participant_id =="P_24_T2"| resp_locked_data$participant_id =="P_25_T1"| resp_locked_data$participant_id =="P_26_T2"|
  resp_locked_data$participant_id =="P_27_T1"| resp_locked_data$participant_id =="P_28_T2"| resp_locked_data$participant_id =="P_29_T2"|
  resp_locked_data$participant_id =="P_30_T1",]$stimulation <- "sham"

resp_locked_data[
  resp_locked_data$participant_id =="C_01_T2"| resp_locked_data$participant_id =="C_02_T2"| resp_locked_data$participant_id =="C_03_T1"|
  resp_locked_data$participant_id =="C_04_T1"| resp_locked_data$participant_id =="C_05_T2"| resp_locked_data$participant_id =="C_06_T1"|
  resp_locked_data$participant_id =="C_07_T2"| resp_locked_data$participant_id =="C_08_T2"| resp_locked_data$participant_id =="C_09_T1"|
  resp_locked_data$participant_id =="C_10_T1"| resp_locked_data$participant_id =="C_11_T1"| resp_locked_data$participant_id =="C_12_T1"|
  resp_locked_data$participant_id =="C_13_T2"| resp_locked_data$participant_id =="C_15_T1"| resp_locked_data$participant_id =="C_16_T2"|
  resp_locked_data$participant_id =="C_17_T1"| resp_locked_data$participant_id =="C_18_T1"| resp_locked_data$participant_id =="C_19_T2"|
  resp_locked_data$participant_id =="C_20_T2"| resp_locked_data$participant_id =="C_21_T1"| resp_locked_data$participant_id =="C_22_T1"|
  resp_locked_data$participant_id =="C_23_T2"| resp_locked_data$participant_id =="C_24_T2"| resp_locked_data$participant_id =="C_25_T2"|
  resp_locked_data$participant_id =="C_26_T2"| resp_locked_data$participant_id =="C_27_T1"| resp_locked_data$participant_id =="C_28_T1"|
  resp_locked_data$participant_id =="C_29_T2"| resp_locked_data$participant_id =="C_30_T1"| resp_locked_data$participant_id =="P_01_T2"|
  resp_locked_data$participant_id =="P_02_T1"| resp_locked_data$participant_id =="P_03_T1"| resp_locked_data$participant_id =="P_04_T2"|
  resp_locked_data$participant_id =="P_05_T2"| resp_locked_data$participant_id =="P_06_T2"| resp_locked_data$participant_id =="P_07_T1"|
  resp_locked_data$participant_id =="P_08_T1"| resp_locked_data$participant_id =="P_09_T1"| resp_locked_data$participant_id =="P_10_T2"|
  resp_locked_data$participant_id =="P_11_T1"| resp_locked_data$participant_id =="P_12_T1"| resp_locked_data$participant_id =="P_13_T2"|
  resp_locked_data$participant_id =="P_15_T2"| resp_locked_data$participant_id =="P_16_T2"| resp_locked_data$participant_id =="P_17_T1"|
  resp_locked_data$participant_id =="P_18_T2"| resp_locked_data$participant_id =="P_19_T2"| resp_locked_data$participant_id =="P_20_T1"|
  resp_locked_data$participant_id =="P_21_T1"| resp_locked_data$participant_id =="P_22_T2"| resp_locked_data$participant_id =="P_23_T2"|
  resp_locked_data$participant_id =="P_24_T1"| resp_locked_data$participant_id =="P_25_T2"| resp_locked_data$participant_id =="P_26_T1"|
  resp_locked_data$participant_id =="P_27_T2"| resp_locked_data$participant_id =="P_28_T1"| resp_locked_data$participant_id =="P_29_T1"|
  resp_locked_data$participant_id =="P_30_T2",]$stimulation <- "verum"



# Add group and session variables and remove string '_T1/T2' from participant ID (to get correct number of factor levels later); make all factors
resp_locked_data <- resp_locked_data  %>% 
  dplyr::mutate(
    group          = factor(ifelse(substr(participant_id, 1,1) == "C", "HC", "OCD")),
    session        = factor(ifelse(substr(participant_id, 6,7) == "T1", "T1", "T2")),
    participant_id = factor(substr(participant_id, 1,4)), 
    condition      = factor(condition, levels = c("incorrect", "correct")),
    stimulation    = factor(stimulation)
    )


# Exclude P_02 (due to retainer) and C_02 (as preregistered: patients are excluded with their match)
resp_locked_data <- resp_locked_data[resp_locked_data$participant_id != "P_02" & resp_locked_data$participant_id != "C_02",]
resp_locked_data$participant_id <- droplevels(resp_locked_data$participant_id)
```

### ERP Trajectories {.tabset}
***

#### Per participant

```{r plot-per-participant-resp-locked, fig.height = 9, fig.width = 10, cache = knitr_cache_enabled}

# Plot ERPs per participant for HC
plot_per_participant_HC <- ggplot(resp_locked_data[resp_locked_data$group == "HC",],aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour = stimulation:condition)) +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-15, 15), xlim=c(-500,600)) +
  scale_y_continuous(breaks=seq(-15, 15, 10)) +
  scale_x_continuous(breaks=c(-200, 100, 400)) +
  scale_colour_manual(values = my_figure_colors) +  
  facet_wrap(.~ participant_id+session, ncol = 8, labeller = labeller(.multi_line = FALSE, sep=": ")) +
  ggtitle("ERPs per participant: HC") +  
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +  
  theme(strip.text.x = element_text(margin = margin(0, 0, 0, 0))) +
  my_figure_theme


# Save plot
ggsave("./figures/figure_ERPs_per_participant_HC.tiff", width = 36, height = 26, units = "cm", dpi=600, compression = "lzw")


# Display plot
plot_per_participant_HC


# Plot ERPs per participant for OCD
plot_per_participant_OCD <- ggplot(resp_locked_data[resp_locked_data$group == "OCD",],aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour = stimulation:condition)) +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-15, 15), xlim=c(-500,600)) +
  scale_y_continuous(breaks=seq(-15, 15, 10)) +
  scale_x_continuous(breaks=c(-200, 100, 400)) +
  scale_colour_manual(values = my_figure_colors) +  
  facet_wrap(.~ participant_id+session, ncol = 8, labeller = labeller(.multi_line = FALSE, sep=": ")) +
  ggtitle("ERPs per participant: OCD") +  
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +  
  theme(strip.text.x = element_text(margin = margin(0, 0, 0, 0))) +
  my_figure_theme


# Save plot
ggsave("./figures/figure_ERPs_per_participant_OCD.tiff", width = 36, height = 26, units = "cm", dpi=600, compression = "lzw")


# Display plot
plot_per_participant_OCD
```
<br><br>

#### ERN/CRN as a function of group and stimulation {.active}

```{r plot-group-by-stimulation-resp-locked, fig.height = 10, fig.width = 10, dpi=600, fig.cap = "Note. Response-locked ERP trajectories (at FCz) as a function of (A) stimulation by group and (B) group by stimulation. Shaded areas represent 95% CIs and are adjusted for within-participant designs as described by Morey (2008).", cache = knitr_cache_enabled}

# Calculate running within-participant CIs
running_CIs_group_stimulation_resp_locked <- resp_locked_data %>%
  split(.$time) %>%
  map(~summarySEwithinO(data = .,
                       measurevar = "FCz",
                       withinvars = c("condition", "stimulation"),
                       betweenvars = "group",
                       idvar = "participant_id"))


# Calculate running within-participant CIs 
CIs_group_stimulation_resp_locked <- purrr::map_df(running_CIs_group_stimulation_resp_locked,magrittr::extract) %>%
  dplyr::mutate(
    time = rep(unique(resp_locked_data$time), each = 8) # Note: 8 refers to the number of conditions (group x stimulation x response_type = 2 x 2 x 2 = 8)
    )


# Plot ERPs per stimulation by group
plot_stimulation_by_group_resp_locked <- ggplot(resp_locked_data,aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked, aes(ymin = FCz-ci, ymax = FCz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-10,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-12,10,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal", "correct: sham", "correct: cathodal")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  ggtitle("ERPs per stimulation by group") +
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +
  my_figure_theme


# Plot ERPs per group by stimulation
plot_group_by_stimulation_resp_locked <- ggplot(resp_locked_data,aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:group)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked, aes(ymin = FCz-ci, ymax = FCz+ci, fill = condition:group), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") + 
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-10,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-12,10,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c( "error: HC", "error: OCD", "correct: HC", "correct: OCD")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ stimulation) + 
  ggtitle("ERPs per group by stimulation") +
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +
  my_figure_theme 


# Arrange plots
figure_ERP_trajectories_resp_locked <- ggdraw() +
  draw_plot(plot_stimulation_by_group_resp_locked,  x = 0, y =.5, width = 1, height =.5) +
  draw_plot(plot_group_by_stimulation_resp_locked,  x = 0, y = 0, width = 1, height =.5) +
  draw_plot_label(c("A", "B"), c(0, 0), c(1, .5), size = 15)


# Save plot
ggsave("./figures/figure_ERP_trajectories.tiff", width = 20, height = 20, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_resp_locked
```
<br>
*Note: CRN is clearly visible but positive. This is quite typical for mastoid reference. *
<br><br>

```{r plot-group-by-stimulation-resp-locked-ERPs-presentation, fig.height = 5, fig.width = 10, dpi=600, fig.cap = "Note. Response-locked ERP trajectories (at FCz) as a function of stimulation by group. Shaded areas represent 95% CIs and are adjusted for within-participant designs as described by Morey (2008).", cache = knitr_cache_enabled, eval = FALSE}


# Plot ERPs per stimulation by group (only sham)
figure_ERP_trajectories_ERN_1 <- ggplot(resp_locked_data[resp_locked_data$condition == "incorrect" & resp_locked_data$stimulation == "sham",],aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked[CIs_group_stimulation_resp_locked$condition == "incorrect" & CIs_group_stimulation_resp_locked$stimulation == "sham",], aes(ymin = FCz-ci, ymax = FCz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-10,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-12,10,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +
  my_figure_theme


# Save plot
#ggsave("./figures/figure_ERP_trajectories_ERN_1.tiff", width = 20, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_ERN_1



# Plot ERPs per stimulation by group
figure_ERP_trajectories_ERN_2 <- ggplot(resp_locked_data[resp_locked_data$condition == "incorrect",],aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked[CIs_group_stimulation_resp_locked$condition == "incorrect",], aes(ymin = FCz-ci, ymax = FCz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-10,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-12,10,2),expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal", "correct: sham", "correct: cathodal")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +
  my_figure_theme


# Save plot
#ggsave("./figures/figure_ERP_trajectories_ERN_2.tiff", width = 20, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_ERN_2




# Plot ERPs per stimulation by group
figure_ERP_trajectories_CRN_1 <- ggplot(resp_locked_data[resp_locked_data$condition == "incorrect" | (resp_locked_data$condition == "correct" & resp_locked_data$stimulation == "sham"),],aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked[CIs_group_stimulation_resp_locked$condition == "incorrect" | (CIs_group_stimulation_resp_locked$condition == "correct" & CIs_group_stimulation_resp_locked$stimulation == "sham"),], aes(ymin = FCz-ci, ymax = FCz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-10,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-12,10,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal", "correct: sham")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +
  my_figure_theme


# Save plot
#ggsave("./figures/figure_ERP_trajectories_CRN_1.tiff", width = 20, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_CRN_1



# Plot ERPs per stimulation by group
figure_ERP_trajectories_CRN_2 <- ggplot(resp_locked_data,aes(time, FCz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked, aes(ymin = FCz-ci, ymax = FCz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=0, xmax=100), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-10,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-12,10,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal", "correct: sham", "correct: cathodal")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  labs(x = "\nTime (ms)", y = expression(paste("FCz (",mu,"V)")), colour = "") +
  my_figure_theme


# Save plot
#ggsave("./figures/figure_ERP_trajectories_CRN_2.tiff", width = 20, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_CRN_2
```
<br><br>

#### Pe as a function of group and stimulation

```{r plot-group-by-stimulation-resp-locked-Pe, fig.height = 5, fig.width = 10, dpi=600, fig.cap = "Note. Response-locked ERP trajectories (at Pz) as a function of (A) stimulation by group and (B) group by stimulation. Shaded areas represent 95% CIs and are adjusted for within-participant designs as described by Morey (2008).", cache = knitr_cache_enabled}

# Calculate running within-participant CIs
running_CIs_group_stimulation_resp_locked_Pe <- resp_locked_data %>%
  split(.$time) %>%
  map(~summarySEwithinO(data = .,
                       measurevar = "Pz",
                       withinvars = c("condition", "stimulation"),
                       betweenvars = "group",
                       idvar = "participant_id"))


# Calculate running within-participant CIs 
CIs_group_stimulation_resp_locked_Pe <- purrr::map_df(running_CIs_group_stimulation_resp_locked_Pe,magrittr::extract) %>%
  dplyr::mutate(
    time = rep(unique(resp_locked_data$time), each = 8) # Note: 8 refers to the number of conditions (group x stimulation x response_type = 2 x 2 x 2 = 8)
    )


# Plot ERPs per stimulation by group
figure_ERP_trajectories_Pe <- ggplot(resp_locked_data[resp_locked_data$condition == "incorrect",],aes(time, Pz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_resp_locked_Pe[CIs_group_stimulation_resp_locked_Pe$condition == "incorrect",], aes(ymin = Pz-ci, ymax = Pz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=200, xmax=400), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-2,12), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-2,12,2),expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  labs(x = "\nTime (ms)", y = expression(paste("Pz (",mu,"V)")), colour = "") +
  my_figure_theme


# Save plot
ggsave("./figures/figure_ERP_trajectories_Pe.tiff", width = 20, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_Pe
```
<br>

### ERP Topographies {.tabset}
***

#### Per group and response type

```{r topographies-resp-locked, fig.height = 7, fig.width = 10, dpi=600, fig.cap = "Note. Response-locked ERP topograhpies (from 0 to 100 ms) separately for both groups and response conditions.", cache = knitr_cache_enabled}

# Convert to long format for electrodes
data_topo <- resp_locked_data %>%
  # Remove channels of no interest
  dplyr::select(-IO1, -M1, -F9, -F10) %>%
  # Change from wide to long format for electrodes
  tidyr::gather(., electrode, amplitude, Fp1:O2, factor_key = TRUE)  %>%
  # Select time windows
  dplyr::filter(time >= 0 & time <= 100) %>%
  # Add electrode information
  electrode_locations(., electrode = "electrode", drop = FALSE, montage = NULL)


# Calculate difference score incorrect-correct
data_topo_diff <- left_join(data_topo[data_topo$condition == "correct",], data_topo[data_topo$condition == "incorrect",], by = c("participant_id", "time", "group","session","stimulation","electrode")) %>%
  # Calculate incorrect - correct
  dplyr::mutate(amplitude = amplitude.y - amplitude.x) %>%
  # Remove unncessary columns
  dplyr::select(-c(condition.x, condition.y, amplitude.x, amplitude.y)) %>%
  # Select time windows
  dplyr::filter(time >= 0 & time <= 100) %>%
  # Add electrode information
  electrode_locations(., electrode = "electrode", drop = FALSE, montage = NULL)


# Calculate difference score sham-verum
data_topo_diff_stimulation <- left_join(data_topo[data_topo$stimulation == "verum",], data_topo[data_topo$stimulation == "sham",], by = c("participant_id", "time", "group","condition","electrode")) %>%
  # Calculate sham - verum
  dplyr::mutate(amplitude = amplitude.y - amplitude.x) %>%
  # Remove unncessary columns
  dplyr::select(-c(stimulation.x, stimulation.y, amplitude.x, amplitude.y)) %>%
  # Select time windows
  dplyr::filter(time >= 0 & time <= 100) %>%
  # Add electrode information
  electrode_locations(., electrode = "electrode", drop = FALSE, montage = NULL)


# Draw topographies
plot_topo <- ggplot(data_topo, aes(x = x, y = y, fill = amplitude, label = electrode)) + 
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.9) + 
    scale_fill_distiller(palette = "RdBu", limits = c(-5, 12)) + 
    theme_void() + 
    coord_equal() + 
    labs(fill = expression(paste(mu, "V"))) + 
    theme(strip.text = element_text(size = 12, hjust = .5, face = "bold"),legend.position = "bottom") + 
    facet_grid(cols = vars(condition), rows = vars(group), switch = "y") +
    theme(strip.text.y.left = element_text(angle = 0)) # to show group labels horizontally


plot_topo_diff <- ggplot(data_topo_diff, aes(x = x, y = y, fill = amplitude, label = electrode)) + 
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.9) + 
    scale_fill_distiller(palette = "RdBu", limits = c(-12, 0)) + 
    theme_void() + 
    coord_equal() + 
    ggtitle("incorrect-correct") +
    labs(fill = expression(paste(mu, "V"))) + 
    theme(plot.title = element_text(size = 12, hjust = .5, face = "bold"),legend.position = "bottom") + 
    facet_grid(rows = vars(group)) +
    theme(strip.text.y = element_blank())  # remove facet labels

    
# Store legends separately in order to center them correctly
legend_topo      <- get_legend(plot_topo)
legend_topo_diff <- get_legend(plot_topo_diff)


# Remove previous legends from plots 
plot_topo       <- plot_topo      + theme(legend.position = "none")
plot_topo_diff  <- plot_topo_diff + theme(legend.position = "none")


# Arrange plots
figure_ERP_topographies <- ggdraw() +
  draw_plot(plot_topo,        x =  0,  y =  .1, width = .66,  height = .90)+
  draw_plot(plot_topo_diff,   x = .66, y =  .1, width = .306, height = .90)+
  draw_plot(legend_topo,      x = .23, y =  0,  width = .2,   height = .1) + 
  draw_plot(legend_topo_diff, x = .7,  y =  0,  width = .2,   height = .1)


# Save plot
ggsave("./figures/figure_ERP_topographies.tiff", width = 16, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_topographies
```
<br><br>

#### Per group, response type, and stimulation

```{r topographies-stimulation-resp-locked, fig.height = 7, fig.width = 12, dpi=600, fig.cap = "Note. Response-locked ERP topograhpies (from 0 to 100 ms) separately for both groups, response conditions, and stimulation conditions.", cache = knitr_cache_enabled}

# Draw topographies including stimulation
plot_topo_stimulation <- plot_topo + 
  facet_grid(cols = vars(condition, stimulation), rows = vars(group), switch = "y") +
  scale_fill_distiller(palette = "RdBu", limits = c(-6, 12)) + 
  theme(legend.position = "bottom") 


plot_topo_diff_stimulation <- ggplot(data_topo_diff_stimulation, aes(x = x, y = y, fill = amplitude, label = electrode)) + 
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.9) + 
    scale_fill_distiller(palette = "RdBu", limits = c(-2, 2)) + 
    theme_void() + 
    coord_equal() + 
    ggtitle("sham-verum") +
    labs(fill = expression(paste(mu, "V"))) + 
    theme(plot.title = element_text(size = 12, vjust = 3, hjust = .5, face = "bold"),legend.position = "bottom") + 
    facet_grid(cols = vars(condition), rows = vars(group)) +
    theme(strip.text.y = element_blank())  # remove facet labels



# Store legends separately in order to center them correctly
legend_topo_stimulation      <- get_legend(plot_topo_stimulation)
legend_topo_diff_stimulation <- get_legend(plot_topo_diff_stimulation)


# Remove previous legends from plots 
plot_topo_stimulation       <- plot_topo_stimulation      + theme(legend.position = "none")
plot_topo_diff_stimulation  <- plot_topo_diff_stimulation + theme(legend.position = "none")


# Arrange plots
figure_ERP_topographies_stimulation <- ggdraw() +
  draw_plot(plot_topo_stimulation,        x =  0,  y =  0,   width = .66,  height = .95)+
  draw_plot(plot_topo_diff_stimulation,   x = .67, y =  0,   width = .306, height = .95)+
  draw_plot(legend_topo_stimulation,      x = .23, y =  .08, width = .2,   height = .1) + 
  draw_plot(legend_topo_diff_stimulation, x = .7,  y =  .08, width = .2,   height = .1)



# Save plot
ggsave("./figures/figure_ERP_topographies_stimulation.tiff", width = 20, height = 14, units = "cm", dpi=600, compression = "lzw")


# Display plot
figure_ERP_topographies_stimulation



# ggplot(data_topo_diff_stimulation[data_topo_diff_stimulation$condition == "incorrect",], aes(x = x, y = y, fill = amplitude, label = electrode)) + 
#     geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.9) + 
#     scale_fill_distiller(palette = "RdBu", limits = c(-1, 1)) + 
#     theme_void() + 
#     coord_equal() + 
#     ggtitle("sham-verum") +
#     labs(fill = expression(paste(mu, "V"))) + 
#     theme(plot.title = element_text(size = 12, vjust = 3, hjust = .5, face = "bold"),legend.position = "bottom") + 
#     facet_grid(rows = vars(group))
# 
# ggplot(data_topo_diff_stimulation[data_topo_diff_stimulation$condition == "correct",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
#     geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.9) +
#     scale_fill_distiller(palette = "RdBu", limits = c(-1.5, 1)) +
#     theme_void() +
#     coord_equal() +
#     ggtitle("sham-verum") +
#     labs(fill = expression(paste(mu, "V"))) +
#     theme(plot.title = element_text(size = 12, vjust = 3, hjust = .5, face = "bold"),legend.position = "bottom") +
#     facet_grid(rows = vars(group))
```
<br><br>

```{r topographies-resp-locked-presentation, fig.height = 5, fig.width = 4, dpi=300, eval = FALSE}

# Plot ERN HC
figure_topo_ERN_HC <- ggplot(data_topo[data_topo$condition == "incorrect" & data_topo$group == "HC",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(-5, 6.2)) +
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_ERN_HC.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")


# Plot ERN OCD
figure_topo_ERN_OCD <- ggplot(data_topo[data_topo$condition == "incorrect" & data_topo$group == "OCD",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(-5, 6.2)) +
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_ERN_OCD.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")



# Plot CRN HC
figure_topo_CRN_HC <- ggplot(data_topo[data_topo$condition == "correct" & data_topo$group == "HC",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(0, 12)) +
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_CRN_HC.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")


# Plot CRN OCD
figure_topo_CRN_OCD <- ggplot(data_topo[data_topo$condition == "correct" & data_topo$group == "OCD",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(0, 12)) +
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_CRN_OCD.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")


##### Pe

data_topo_Pe <- resp_locked_data %>%
  # Remove channels of no interest
  dplyr::select(-IO1, -M1, -F9, -F10) %>%
  # Change from wide to long format for electrodes
  tidyr::gather(., electrode, amplitude, Fp1:O2, factor_key = TRUE)  %>%
  # Select time windows
  dplyr::filter(time >= 200 & time <= 400) %>%
  # Add electrode information
  electrode_locations(., electrode = "electrode", drop = FALSE, montage = NULL)


# Plot Pe HC
figure_topo_Pe_HC <- ggplot(data_topo_Pe[data_topo_Pe$condition == "incorrect" & data_topo_Pe$group == "HC",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(2, 10)) + 
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_Pe_HC.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")


# Plot Pe OCD
figure_topo_Pe_OCD <- ggplot(data_topo_Pe[data_topo_Pe$condition == "incorrect" & data_topo_Pe$group == "OCD",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(2, 10)) +
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_Pe_OCD.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")
```
<br><br>

## Stimulus-locked ERPs
***

<!-- Load and clean data -->

```{r load-and-clean-data-stim-locked}

# Load data
stim_locked_data <- read.csv(file = "./data/stimulus_locked_data_for_plots.csv", header = TRUE)


# Rename conditions
stim_locked_data$condition[stim_locked_data$condition == 9]  <- "correct";
stim_locked_data$condition[stim_locked_data$condition == 10] <- "incorrect";



# Add stimulation condition variable
stim_locked_data$stimulation <- NA
stim_locked_data[
  stim_locked_data$participant_id =="C_01_T1"| stim_locked_data$participant_id =="C_02_T1"| stim_locked_data$participant_id =="C_03_T2"|
  stim_locked_data$participant_id =="C_04_T2"| stim_locked_data$participant_id =="C_05_T1"| stim_locked_data$participant_id =="C_06_T2"|
  stim_locked_data$participant_id =="C_07_T1"| stim_locked_data$participant_id =="C_08_T1"| stim_locked_data$participant_id =="C_09_T2"|
  stim_locked_data$participant_id =="C_10_T2"| stim_locked_data$participant_id =="C_11_T2"| stim_locked_data$participant_id =="C_12_T2"|
  stim_locked_data$participant_id =="C_13_T1"| stim_locked_data$participant_id =="C_15_T2"| stim_locked_data$participant_id =="C_16_T1"|
  stim_locked_data$participant_id =="C_17_T2"| stim_locked_data$participant_id =="C_18_T2"| stim_locked_data$participant_id =="C_19_T1"|
  stim_locked_data$participant_id =="C_20_T1"| stim_locked_data$participant_id =="C_21_T2"| stim_locked_data$participant_id =="C_22_T2"|
  stim_locked_data$participant_id =="C_23_T1"| stim_locked_data$participant_id =="C_24_T1"| stim_locked_data$participant_id =="C_25_T1"|
  stim_locked_data$participant_id =="C_26_T1"| stim_locked_data$participant_id =="C_27_T2"| stim_locked_data$participant_id =="C_28_T2"|
  stim_locked_data$participant_id =="C_29_T1"| stim_locked_data$participant_id =="C_30_T2"| stim_locked_data$participant_id =="P_01_T1"|
  stim_locked_data$participant_id =="P_02_T2"| stim_locked_data$participant_id =="P_03_T2"| stim_locked_data$participant_id =="P_04_T1"|
  stim_locked_data$participant_id =="P_05_T1"| stim_locked_data$participant_id =="P_06_T1"| stim_locked_data$participant_id =="P_07_T2"|
  stim_locked_data$participant_id =="P_08_T2"| stim_locked_data$participant_id =="P_09_T2"| stim_locked_data$participant_id =="P_10_T1"|
  stim_locked_data$participant_id =="P_11_T2"| stim_locked_data$participant_id =="P_12_T2"| stim_locked_data$participant_id =="P_13_T1"|
  stim_locked_data$participant_id =="P_15_T1"| stim_locked_data$participant_id =="P_16_T1"| stim_locked_data$participant_id =="P_17_T2"|
  stim_locked_data$participant_id =="P_18_T1"| stim_locked_data$participant_id =="P_19_T1"| stim_locked_data$participant_id =="P_20_T2"|
  stim_locked_data$participant_id =="P_21_T2"| stim_locked_data$participant_id =="P_22_T1"| stim_locked_data$participant_id =="P_23_T1"|
  stim_locked_data$participant_id =="P_24_T2"| stim_locked_data$participant_id =="P_25_T1"| stim_locked_data$participant_id =="P_26_T2"|
  stim_locked_data$participant_id =="P_27_T1"| stim_locked_data$participant_id =="P_28_T2"| stim_locked_data$participant_id =="P_29_T2"|
  stim_locked_data$participant_id =="P_30_T1",]$stimulation <- "sham"

stim_locked_data[
  stim_locked_data$participant_id =="C_01_T2"| stim_locked_data$participant_id =="C_02_T2"| stim_locked_data$participant_id =="C_03_T1"|
  stim_locked_data$participant_id =="C_04_T1"| stim_locked_data$participant_id =="C_05_T2"| stim_locked_data$participant_id =="C_06_T1"|
  stim_locked_data$participant_id =="C_07_T2"| stim_locked_data$participant_id =="C_08_T2"| stim_locked_data$participant_id =="C_09_T1"|
  stim_locked_data$participant_id =="C_10_T1"| stim_locked_data$participant_id =="C_11_T1"| stim_locked_data$participant_id =="C_12_T1"|
  stim_locked_data$participant_id =="C_13_T2"| stim_locked_data$participant_id =="C_15_T1"| stim_locked_data$participant_id =="C_16_T2"|
  stim_locked_data$participant_id =="C_17_T1"| stim_locked_data$participant_id =="C_18_T1"| stim_locked_data$participant_id =="C_19_T2"|
  stim_locked_data$participant_id =="C_20_T2"| stim_locked_data$participant_id =="C_21_T1"| stim_locked_data$participant_id =="C_22_T1"|
  stim_locked_data$participant_id =="C_23_T2"| stim_locked_data$participant_id =="C_24_T2"| stim_locked_data$participant_id =="C_25_T2"|
  stim_locked_data$participant_id =="C_26_T2"| stim_locked_data$participant_id =="C_27_T1"| stim_locked_data$participant_id =="C_28_T1"|
  stim_locked_data$participant_id =="C_29_T2"| stim_locked_data$participant_id =="C_30_T1"| stim_locked_data$participant_id =="P_01_T2"|
  stim_locked_data$participant_id =="P_02_T1"| stim_locked_data$participant_id =="P_03_T1"| stim_locked_data$participant_id =="P_04_T2"|
  stim_locked_data$participant_id =="P_05_T2"| stim_locked_data$participant_id =="P_06_T2"| stim_locked_data$participant_id =="P_07_T1"|
  stim_locked_data$participant_id =="P_08_T1"| stim_locked_data$participant_id =="P_09_T1"| stim_locked_data$participant_id =="P_10_T2"|
  stim_locked_data$participant_id =="P_11_T1"| stim_locked_data$participant_id =="P_12_T1"| stim_locked_data$participant_id =="P_13_T2"|
  stim_locked_data$participant_id =="P_15_T2"| stim_locked_data$participant_id =="P_16_T2"| stim_locked_data$participant_id =="P_17_T1"|
  stim_locked_data$participant_id =="P_18_T2"| stim_locked_data$participant_id =="P_19_T2"| stim_locked_data$participant_id =="P_20_T1"|
  stim_locked_data$participant_id =="P_21_T1"| stim_locked_data$participant_id =="P_22_T2"| stim_locked_data$participant_id =="P_23_T2"|
  stim_locked_data$participant_id =="P_24_T1"| stim_locked_data$participant_id =="P_25_T2"| stim_locked_data$participant_id =="P_26_T1"|
  stim_locked_data$participant_id =="P_27_T2"| stim_locked_data$participant_id =="P_28_T1"| stim_locked_data$participant_id =="P_29_T1"|
  stim_locked_data$participant_id =="P_30_T2",]$stimulation <- "verum"



# Add group and session variables and remove string '_T1/T2' from participant ID (to get correct number of factor levels later); make all factors
stim_locked_data <- stim_locked_data  %>% 
  dplyr::mutate(
    group          = factor(ifelse(substr(participant_id, 1,1) == "C", "HC", "OCD")),
    session        = factor(ifelse(substr(participant_id, 6,7) == "T1", "T1", "T2")),
    participant_id = factor(substr(participant_id, 1,4)), 
    condition      = factor(condition, levels = c("incorrect","correct")),
    stimulation    = factor(stimulation)
    )


# Exclude P_02 (due to retainer) and C_02 (as preregistered: patients are excluded with their match)
stim_locked_data <- stim_locked_data[stim_locked_data$participant_id != "P_02" & stim_locked_data$participant_id != "C_02",]
stim_locked_data$participant_id <- droplevels(stim_locked_data$participant_id)
```

### ERP Trajectories {.tabset}
***

#### ERPs as a function of group and stimulation {.active}

```{r plot-group-by-stimulation-stim-locked, fig.height = 10, fig.width = 10, dpi=600, fig.cap = "Note. Stimulus-locked ERP trajectories (at CPz, but note that for N2 FCz should better be displayed) as a function of (A) stimulation by group and (B) group by stimulation. Shaded areas represent 95% CIs and are adjusted for within-participant designs as described by Morey (2008).", cache = knitr_cache_enabled}

# Calculate running within-participant CIs
running_CIs_group_stimulation_stim_locked <- stim_locked_data %>%
  split(.$time) %>%
  map(~summarySEwithinO(data = .,
                       measurevar = "CPz",
                       withinvars = c("condition", "stimulation"),
                       betweenvars = "group",
                       idvar = "participant_id"))


# Calculate running within-participant CIs 
CIs_group_stimulation_stim_locked <- purrr::map_df(running_CIs_group_stimulation_stim_locked,magrittr::extract) %>%
  dplyr::mutate(
    time = rep(unique(stim_locked_data$time), each = 8) # Note: 8 refers to the number of conditions (group x stimulation x response_type = 2 x 2 x 2 = 8)
    )


# Plot ERPs per stimulation by group
plot_stimulation_by_group_stim_locked <- ggplot(stim_locked_data,aes(time, CPz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_stim_locked, aes(ymin = CPz-ci, ymax = CPz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=200, xmax=295), alpha=.2, fill="grey70") +
  geom_ribbon(aes(xmin=305, xmax=500), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-2,14), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-2,14,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal", "correct: sham", "correct: cathodal")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  ggtitle("ERPs per stimulation by group") +
  labs(x = "\nTime (ms)", y = expression(paste("CPz (",mu,"V)")), colour = "") +
  my_figure_theme


# Plot ERPs per group by stimulation
plot_group_by_stimulation_stim_locked <- ggplot(stim_locked_data,aes(time, CPz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:group)) +
  geom_ribbon(data = CIs_group_stimulation_stim_locked, aes(ymin = CPz-ci, ymax = CPz+ci, fill = condition:group), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") + 
  geom_ribbon(aes(xmin=200, xmax=295), alpha=.2, fill="grey70") +
  geom_ribbon(aes(xmin=305, xmax=500), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-2,14), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-2,14,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c( "error: HC", "error: OCD", "correct: HC", "correct: OCD")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ stimulation) + 
  ggtitle("ERPs per group by stimulation") +
  labs(x = "\nTime (ms)", y = expression(paste("CPz (",mu,"V)")), colour = "") +
  my_figure_theme 


# Arrange plots
figure_ERP_trajectories_stim_locked <- ggdraw() +
  draw_plot(plot_stimulation_by_group_stim_locked,  x = 0, y =.5, width = 1, height =.5) +
  draw_plot(plot_group_by_stimulation_stim_locked,  x = 0, y = 0, width = 1, height =.5) +
  draw_plot_label(c("A", "B"), c(0, 0), c(1, .5), size = 15)


# Save plot
ggsave("./figures/figure_ERP_trajectories_stim_locked.tiff", width = 20, height = 20, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_stim_locked
```
<br>

```{r plot-group-by-stimulation-stim-locked-only-presentation, fig.height = 5, fig.width = 10, dpi=600, fig.cap = "Note. Stimulus-locked ERP trajectories (at CPz, but note that for N2 FCz should better be displayed) as a function of (A) stimulation by group and (B) group by stimulation. Shaded areas represent 95% CIs and are adjusted for within-participant designs as described by Morey (2008).", cache = knitr_cache_enabled, eval = FALSE}


# Plot ERPs per stimulation by group
figure_ERP_trajectories_stim_locked <- ggplot(stim_locked_data,aes(time, CPz)) +
  stat_summary(fun = mean, geom = "line", size = 0.5, linetype = "solid", aes(colour= condition:stimulation)) +
  geom_ribbon(data = CIs_group_stimulation_stim_locked, aes(ymin = CPz-ci, ymax = CPz+ci, fill = condition:stimulation), alpha = 0.15) + guides(fill = "none") +
  geom_vline(xintercept = 0, linetype = "dashed", colour="grey" ) +
  geom_hline(yintercept = 0, linetype = "dashed", colour="grey") +  
  geom_ribbon(aes(xmin=300, xmax=500), alpha=.2, fill="grey70") +
  coord_cartesian(ylim=c(-2,14), xlim=c(-500,1000)) +
  scale_y_continuous(breaks=seq(-2,14,2), expand = c(0, 0)) +
  scale_x_continuous(breaks=seq(-500,1000,200), expand = c(0, 0)) +
  scale_colour_manual(values = my_figure_colors, labels = c("error: sham", "error: cathodal", "correct: sham", "correct: cathodal")) +
  scale_fill_manual(values = my_figure_colors) +
  facet_grid(. ~ group) + 
  labs(x = "\nTime (ms)", y = expression(paste("CPz (",mu,"V)")), colour = "") +
  my_figure_theme


# Save plot
# ggsave("./figures/figure_ERP_trajectories_stim_locked.tiff", width = 20, height = 10, units = "cm", dpi=300, compression = "lzw")


# Display plot
figure_ERP_trajectories_stim_locked
```
<br>

```{r topographies-stim-locked-presentation, fig.height = 5, fig.width = 4, dpi=300, eval = FALSE}

##### P300

data_topo_P300 <- stim_locked_data %>%
  # Remove channels of no interest
  dplyr::select(-IO1, -M1, -F9, -F10) %>%
  # Change from wide to long format for electrodes
  tidyr::gather(., electrode, amplitude, Fp1:O2, factor_key = TRUE)  %>%
  # Select time windows
  dplyr::filter(time >= 300 & time <= 500) %>%
  # Add electrode information
  electrode_locations(., electrode = "electrode", drop = FALSE, montage = NULL)


# Plot P300 HC
figure_topo_P300_HC <- ggplot(data_topo_P300[data_topo_P300$group == "HC",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(0, 8.7)) + 
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_P300_HC.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")


# Plot P300 OCD
figure_topo_P300_OCD <- ggplot(data_topo_P300[data_topo_P300$group == "OCD",], aes(x = x, y = y, fill = amplitude, label = electrode)) +
    geom_topo(grid_res = 300, interp_limit = "head", chan_markers = "point", chan_size = 0.5, head_size = 0.5) +
    scale_fill_distiller(palette = "RdBu", limits = c(0, 8.7)) +
    theme_void() +
    coord_equal() +
    labs(fill = expression(paste(mu, "V"))) +
    theme(strip.text = element_text(size = 18, hjust = .5, face = "bold"),legend.position = "bottom") 

# Save plot
#ggsave("./figures/figure_topo_P300_OCD.tiff", width = 5, height = 6, units = "cm", dpi=300, compression = "lzw")
```
